Add gRPC-free LinuxCaptureServiceBase
With this change, we extract a gRPC-free LinuxCaptureServiceBase from
LinuxCaptureService which holds the common functionality shared by
native Orbit Linux capture service and the cloud collector.

Bug: http://b/216085824
