+++ src/LinuxCaptureService/include/LinuxCaptureService/LinuxCaptureService.h
@@ -5,54 +5,22 @@
 #ifndef LINUX_CAPTURE_SERVICE_LINUX_CAPTURE_SERVICE_H_
 #define LINUX_CAPTURE_SERVICE_LINUX_CAPTURE_SERVICE_H_
 
-#include <absl/container/flat_hash_set.h>
 #include <grpcpp/grpcpp.h>
 
-#include <atomic>
-#include <memory>
-#include <thread>
-
-#include "CaptureServiceBase/CaptureServiceBase.h"
-#include "CaptureServiceBase/StartStopCaptureRequestWaiter.h"
 #include "GrpcProtos/services.grpc.pb.h"
 #include "GrpcProtos/services.pb.h"
-#include "OrbitBase/Logging.h"
-#include "OrbitBase/Profiling.h"
-#include "UserSpaceInstrumentation/InstrumentProcess.h"
+#include "LinuxCaptureService/LinuxCaptureServiceBase.h"
 
 namespace orbit_linux_capture_service {
 
 // Linux implementation of the grpc capture service.
-class LinuxCaptureService final : public orbit_capture_service_base::CaptureServiceBase,
+class LinuxCaptureService final : public LinuxCaptureServiceBase,
                                   public orbit_grpc_protos::CaptureService::Service {
  public:
-  LinuxCaptureService() {
-    instrumentation_manager_ = orbit_user_space_instrumentation::InstrumentationManager::Create();
-  }
-
-  ~LinuxCaptureService() override {
-    if (wait_for_stop_capture_request_thread_.joinable()) {
-      wait_for_stop_capture_request_thread_.join();
-    }
-  }
-
   grpc::Status Capture(
       grpc::ServerContext* context,
       grpc::ServerReaderWriter<orbit_grpc_protos::CaptureResponse,
                                orbit_grpc_protos::CaptureRequest>* reader_writer) override;
-
- private:
-  std::unique_ptr<orbit_user_space_instrumentation::InstrumentationManager>
-      instrumentation_manager_;
-
-  // This method returns when the first of the following happens:
-  // - The client calls WritesDone on reader_writer (similar to
-  //   CaptureService::WaitForStopCaptureRequestFromClient);
-  // - MemoryWatchdog calls OnThresholdExceeded.
-  [[nodiscard]] StopCaptureReason WaitForStopCaptureRequestOrMemoryThresholdExceeded(
-      const std::shared_ptr<orbit_capture_service_base::StartStopCaptureRequestWaiter>&
-          start_stop_capture_request_waiter);
-  std::thread wait_for_stop_capture_request_thread_;
 };
 
 }  // namespace orbit_linux_capture_service
